release 0.1.1 alpha work in progress

+++ scripts/copy-to-github.sh
@@ -1,7 +1,20 @@
 #!/usr/bin/env bash
 # copy files to github for release
+RELEASE=$1
+if [[ "$RELEASE" == "" ]]; then
+    echo "need to specify release"
+    exit 1
+fi  
 export BASE=$(git rev-parse --show-toplevel)
 GITHUB=$HOME/extgit/fanling
+if [[ -f $GITHUB ]]; then
+    rm $GITHUB
+    echo "deleted file" $GITHUB
+fi
+if [[ ! -d $GITHUB ]]; then
+    echo "no directory" $GITHUB
+    exit 1
+fi
 cd $BASE
 rm -rf $BASE/fanling-engine/testfiles*
 rm -rf $BASE/fanling-engine/"??"
@@ -16,6 +29,12 @@ for D in $DD
 do
     cp -R $D $GITHUB
 done
+git commit -am "release $RELEASE"
+git tag $RELEASE
 
 cd $GITHUB
 git add $FF $DD
+git commit -am "release $RELEASE"
+git tag $RELEASE
+
+
